requirements.txt
pytablewriter==1.0.0
# deepchem[torch]==2.7.1
deepchem[torch]==2.7.2.dev20230803163359
numpy==1.22.3
selfies==2.1.1
deepsmiles==1.0.1
mhfp==1.9.6
smizip==1.0.1
brotli==1.1.0
python-snappy==0.6.1
lz4==4.3.3